main: alternate method to get boardname for Android
Currently powertop uses the /etc/boardname to retrive the boardname
as well as set the global_fixed_parameters and global_power_override
global variables. /etc/boardname is a meego specific feature and not
present in Android. This patch enables powertop to get the board name using
getprop system call in case of Android devices.